backend/requirements.txt
django==5.2.14
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
pymongo==4.17.0
dnspython==2.8.0
django-cors-headers==4.9.0
python-dotenv==1.2.2
razorpay==2.0.1
Pillow==12.2.0
django-filter==25.2
